+++ Core/UserInformation/Models/Project.cs
@@ -19,7 +19,7 @@ namespace Unity.Cloud.Identity
 
         readonly IEntityRoleProvider m_EntityRoleProvider;
 
-        readonly GetRequestResponseCache<RangeResultsJson<ProjectMemberInfoJson>> m_GetRequestResponseCache;
+        readonly GetRequestResponseCache<RangeResultsJson<MemberInfoJson>> m_GetRequestResponseCache;
 
         internal Project(ProjectJson projectJson, IServiceHttpClient serviceHttpClient, IServiceHostResolver serviceHostResolver, IEntityRoleProvider entityRoleProvider)
         {
@@ -37,7 +37,7 @@ namespace Unity.Cloud.Identity
 
             m_EntityRoleProvider = entityRoleProvider;
 
-            m_GetRequestResponseCache = new GetRequestResponseCache<RangeResultsJson<ProjectMemberInfoJson>>(60);
+            m_GetRequestResponseCache = new GetRequestResponseCache<RangeResultsJson<MemberInfoJson>>(60);
         }
 
         /// <inheritdoc/>
@@ -76,8 +76,9 @@ namespace Unity.Cloud.Identity
         /// <inheritdoc />
         public async IAsyncEnumerable<IMemberInfo> ListMembersAsync(Range range, [EnumeratorCancellation] CancellationToken cancellationToken = default)
         {
-            var rangeRequest = new RangeRequest<ProjectMemberInfoJson>(GetProjectMembers, 1000);
-            var requestBasePath = $"api/unity/legacy/v1/projects/{Descriptor.ProjectId}/users";
+            var rangeRequest = new RangeRequest<MemberInfoJson>(GetProjectMembers, 1000);
+            // JSON is different here from previous endpoint
+            var requestBasePath = $"api/access/legacy/v1/projects/{Descriptor.ProjectId}/members";
             var results = rangeRequest.Execute(requestBasePath, range, cancellationToken);
             await foreach (var member in results)
             {
@@ -85,39 +86,19 @@ namespace Unity.Cloud.Identity
             }
         }
 
-        async Task<RangeResultsJson<ProjectMemberInfoJson>> GetProjectMembers(string rangeRequestPath, CancellationToken cancellationToken)
+        async Task<RangeResultsJson<MemberInfoJson>> GetProjectMembers(string rangeRequestPath, CancellationToken cancellationToken)
         {
-#if EXPERIMENTAL_WEBGL_PROXY
-            var url = m_ServiceHostResolver.GetResolvedRequestUri("/app-linking/v1alpha1/core");
-
-            if (m_GetRequestResponseCache.TryGetRequestResponseFromCache(rangeRequestPath, out RangeResultsJson<ProjectMemberInfoJson> value))
-            {
-                return value;
-            }
-
-            var coreApiRequest = new CoreApiRequestParams
-            {
-                Path = rangeRequestPath,
-                Method = "Get",
-            };
-            var content = new StringContent(JsonSerialization.Serialize(coreApiRequest), Encoding.UTF8, "application/json");
-            var response = await m_ServiceHttpClient.PostAsync(url, content, cancellationToken: cancellationToken);
-
-            var deserializedResponse = await response.JsonDeserializeAsync<RangeResultsJson<ProjectMemberInfoJson>>();
-            return m_GetRequestResponseCache.AddGetRequestResponseToCache(rangeRequestPath, deserializedResponse);
-#else
             var internalServiceHostResolver = m_ServiceHostResolver.CreateCopyWithDomainResolverOverride(new UnityServicesDomainResolver(true));
             var url = internalServiceHostResolver.GetResolvedRequestUri($"/{rangeRequestPath}");
 
-            if (m_GetRequestResponseCache.TryGetRequestResponseFromCache(url, out RangeResultsJson<ProjectMemberInfoJson> value))
+            if (m_GetRequestResponseCache.TryGetRequestResponseFromCache(url, out RangeResultsJson<MemberInfoJson> value))
             {
                 return value;
             }
 
             var response = await m_ServiceHttpClient.GetAsync(url, cancellationToken: cancellationToken);
-            var deserializedResponse = await response.JsonDeserializeAsync<RangeResultsJson<ProjectMemberInfoJson>>();
+            var deserializedResponse = await response.JsonDeserializeAsync<RangeResultsJson<MemberInfoJson>>();
             return m_GetRequestResponseCache.AddGetRequestResponseToCache(url, deserializedResponse);
-#endif
         }
 
     }