+++ Core/UserInformation/UnityUserInfoJsonProvider.cs
@@ -13,13 +13,15 @@ namespace Unity.Cloud.Identity
 
     internal class UnityUserInfoJsonProvider : IUnityUserInfoJsonProvider
     {
+        readonly string m_UserId;
         readonly IServiceHostResolver m_ServiceHostResolver;
         readonly IServiceHttpClient m_ServiceHttpClient;
 
         readonly GetRequestResponseCache<UnityUserInfoJson> m_GetUnityUserOrganizationRequestResponseCache;
 
-        public UnityUserInfoJsonProvider(IServiceHttpClient serviceHttpClient, IServiceHostResolver serviceHostResolver)
+        public UnityUserInfoJsonProvider(string userId, IServiceHttpClient serviceHttpClient, IServiceHostResolver serviceHostResolver)
         {
+            m_UserId = userId;
             m_ServiceHostResolver = serviceHostResolver;
             m_ServiceHttpClient = serviceHttpClient;
 
@@ -28,31 +30,8 @@ namespace Unity.Cloud.Identity
 
         public async Task<UnityUserInfoJson> GetUnityUserInfoJsonAsync()
         {
-#if EXPERIMENTAL_WEBGL_PROXY
-            var coreApiRequestPath = "api/unity/v1/users/me/organizations";
-            var url = m_ServiceHostResolver.GetResolvedRequestUri("/app-linking/v1alpha1/core");
-            UnityUserInfoJson userInfoJson;
-            if (m_GetUnityUserOrganizationRequestResponseCache.TryGetRequestResponseFromCache(coreApiRequestPath, out UnityUserInfoJson value))
-            {
-                userInfoJson = value;
-            }
-            else
-            {
-                var coreApiRequest = new CoreApiRequestParams
-                {
-                    Path = coreApiRequestPath,
-                    Method = "Get",
-                };
-                var contentBody = new StringContent(JsonSerialization.Serialize(coreApiRequest), Encoding.UTF8, "application/json");
-                var response = await m_ServiceHttpClient.PostAsync(url, contentBody);
-
-                var deserializedResponse = await response.JsonDeserializeAsync<UnityUserInfoJson>();
-                userInfoJson = m_GetUnityUserOrganizationRequestResponseCache.AddGetRequestResponseToCache(coreApiRequestPath, deserializedResponse);
-            }
-            return userInfoJson;
-#else
             var internalServiceHostResolver = m_ServiceHostResolver.CreateCopyWithDomainResolverOverride(new UnityServicesDomainResolver(true));
-            var url = internalServiceHostResolver.GetResolvedRequestUri("/api/unity/v1/users/me/organizations");
+            var url = internalServiceHostResolver.GetResolvedRequestUri($"/api/unity/legacy/v1/users/{m_UserId}/organizations");
             UnityUserInfoJson userInfoJson;
             if (m_GetUnityUserOrganizationRequestResponseCache.TryGetRequestResponseFromCache(url, out UnityUserInfoJson value))
             {
@@ -65,7 +44,6 @@ namespace Unity.Cloud.Identity
                 userInfoJson = m_GetUnityUserOrganizationRequestResponseCache.AddGetRequestResponseToCache(url, deserializedResponse);
             }
             return userInfoJson;
-#endif
         }
     }