+++ Editor/UnityEditorServiceAuthorizer.cs
@@ -56,6 +56,9 @@ namespace Unity.Cloud.Identity.Editor
 
         const double k_ExchangeRequestRetryDelayInSeconds = 0.5;
 
+        const string k_UnityHubUriScheme = "unityhub://";
+        const string k_UnityHubLoginDomain = "login";
+
         IAccessTokenExchanger<TargetClientIdToken, UnityServicesToken>
             m_TargetClientIdTokenToUnityServicesTokenExchanger;
 
@@ -66,25 +69,36 @@ namespace Unity.Cloud.Identity.Editor
         IUnityUserInfoJsonProvider m_UnityUserInfoJsonProvider;
         IGuestProjectJsonProvider m_GuestProjectJsonProvider;
         IOrganizationJsonProvider m_OrganizationJsonProvider;
+        IJwtDecoder m_JwtDecoder;
 
         bool m_UseOverride = false;
         Task<string> m_GetAccessTokenTask;
+
+        DateTime? m_TokenExpiry;
+        bool m_AwaitingExchangeOperation;
 
         internal void OverrideUnityEditorServiceAuthorizer(
             IAccessTokenExchanger<TargetClientIdToken, UnityServicesToken> accessTokenExchanger,
             IUnityEditorAccessTokenProvider unityEditorAccessTokenProvider,
             IUnityUserInfoJsonProvider unityUserInfoJsonProvider = null,
             IGuestProjectJsonProvider guestProjectJsonProvider = null,
-            IOrganizationJsonProvider organizationJsonProvider = null)
+            IOrganizationJsonProvider organizationJsonProvider = null,
+            IJwtDecoder jwtDecoder = null)
         {
             m_TargetClientIdTokenToUnityServicesTokenExchanger = accessTokenExchanger;
             m_UnityEditorAccessTokenProvider = unityEditorAccessTokenProvider;
             m_UnityUserInfoJsonProvider = unityUserInfoJsonProvider;
             m_GuestProjectJsonProvider = guestProjectJsonProvider;
             m_OrganizationJsonProvider = organizationJsonProvider;
+            m_JwtDecoder = jwtDecoder;
 
             m_UseOverride = true;
             InitAuthenticatedUserSession();
+        }
+
+        void OnRefreshAccessToken(bool isRefreshed)
+        {
+            Debug.Log($"Cloud project settings access token refreshed.");
         }
 
         void OnEnable()
@@ -110,9 +124,16 @@ namespace Unity.Cloud.Identity.Editor
             var httpClient = new UnityHttpClient();
             var serviceHostResolver = UnityRuntimeServiceHostResolverFactory.Create();
 
+            m_JwtDecoder ??= new JwtDecoder();
+            if (string.IsNullOrEmpty(UnityServicesToken)) return;
+            var jwt = m_JwtDecoder.Decode(UnityServicesToken);
+            var tokenExpiry = jwt.exp;
+            var userId = jwt.sub;
+            m_TokenExpiry = ConvertTimestamp(tokenExpiry);
+
             if (m_UseOverride)
             {
-                m_AuthenticatedUserSession = new AuthenticatedUserSession(
+                m_AuthenticatedUserSession = new AuthenticatedUserSession(userId,
                     new ServiceHttpClient(httpClient, this, playerSettings),
                     serviceHostResolver,
                     m_UnityUserInfoJsonProvider,
@@ -122,8 +143,7 @@ namespace Unity.Cloud.Identity.Editor
             }
             else
             {
-                m_TargetClientIdTokenToUnityServicesTokenExchanger = new TargetClientIdTokenToUnityServicesTokenExchanger(httpClient, serviceHostResolver);
-                m_AuthenticatedUserSession = new AuthenticatedUserSession(new ServiceHttpClient(httpClient, this, playerSettings), serviceHostResolver);
+                m_AuthenticatedUserSession = new AuthenticatedUserSession(userId, new ServiceHttpClient(httpClient, this, playerSettings), serviceHostResolver);
             }
         }
 
@@ -138,14 +158,12 @@ namespace Unity.Cloud.Identity.Editor
             if (editorHasToken && AccessToken != accessToken)
             {
                 AccessToken = accessToken;
-                UnityServicesToken = string.Empty;
                 AuthenticationState = AuthenticationState.AwaitingLogin;
             }
-
             if (AuthenticationState.Equals(AuthenticationState.LoggedIn) && !editorHasToken)
             {
                 UnityServicesToken = string.Empty;
-
+                m_TokenExpiry = null;
                 AuthenticationState = AuthenticationState.LoggedOut;
             }
             else if (AuthenticationState.Equals(AuthenticationState.LoggedOut) && editorHasToken)
@@ -154,20 +172,56 @@ namespace Unity.Cloud.Identity.Editor
             }
             else if (AuthenticationState.Equals(AuthenticationState.AwaitingLogin) && editorHasToken)
             {
-                // Throttle request to retry token exchange until connectivity is restored
-                if (EditorApplication.timeSinceStartup - m_LastExchangeRequestCheck < k_ExchangeRequestRetryDelayInSeconds)
-                    return;
+                await RefreshUnityTokenFromAccessTokenAsync();
+            }
+        }
 
-                m_LastExchangeRequestCheck = EditorApplication.timeSinceStartup;
+
+        async Task RefreshUnityTokenFromAccessTokenAsync()
+        {
+            // Throttle request to retry token exchange until connectivity is restored
+            if (EditorApplication.timeSinceStartup - m_LastExchangeRequestCheck < k_ExchangeRequestRetryDelayInSeconds)
+                return;
+
+            m_LastExchangeRequestCheck = EditorApplication.timeSinceStartup;
+            if (!m_AwaitingExchangeOperation)
+            {
+                m_AwaitingExchangeOperation = true;
                 try
                 {
-                    await RefreshUnityTokenAsync(accessToken);
+                    Debug.Log("Try refresh the unity services token.");
+                    await RefreshUnityTokenAsync(AccessToken);
                     if (!string.IsNullOrEmpty(UnityServicesToken))
+                    {
+                        if (!m_UseOverride)
+                        {
+                            var jwt = new JwtDecoder().Decode(UnityServicesToken);
+                            var userId = jwt.sub;
+                            var tokenExpiry = jwt.exp;
+                            m_TokenExpiry = ConvertTimestamp(tokenExpiry);
+
+                            var playerSettings = UnityCloudPlayerSettings.Instance;
+                            var httpClient = new UnityHttpClient();
+                            var serviceHostResolver = UnityRuntimeServiceHostResolverFactory.Create();
+
+                            Debug.Log($"Unity services token refreshed.");
+
+                            m_AuthenticatedUserSession =
+                                new AuthenticatedUserSession(userId,
+                                    new ServiceHttpClient(httpClient, this, playerSettings),
+                                    serviceHostResolver);
+                        }
                         AuthenticationState = AuthenticationState.LoggedIn;
+                    }
+
                 }
                 catch (HttpRequestException)
                 {
                     /* silent fail */
+                }
+                finally
+                {
+                    m_AwaitingExchangeOperation = false;
                 }
             }
         }
@@ -201,6 +255,12 @@ namespace Unity.Cloud.Identity.Editor
         {
             if (m_TargetClientIdTokenToUnityServicesTokenExchanger == null)
             {
+                if (!m_UseOverride)
+                {
+                    var httpClient = new UnityHttpClient();
+                    var serviceHostResolver = UnityRuntimeServiceHostResolverFactory.Create();
+                    m_TargetClientIdTokenToUnityServicesTokenExchanger = new TargetClientIdTokenToUnityServicesTokenExchanger(httpClient, serviceHostResolver);
+                }
                 InitAuthenticatedUserSession();
             }
 
@@ -208,14 +268,20 @@ namespace Unity.Cloud.Identity.Editor
             var exchangedToken = await m_TargetClientIdTokenToUnityServicesTokenExchanger.ExchangeAsync(targetClientIdToken);
 
             UnityServicesToken = exchangedToken.AccessToken;
-            if (!string.IsNullOrEmpty(UnityServicesToken))
-                AccessToken = accessToken;
         }
 
         /// <inheritdoc cref="IServiceAuthorizer.AddAuthorization"/>
         public async Task AddAuthorization(HttpHeaders headers)
         {
 #if UNITY_EDITOR
+            ValidateOrRevokeUnityServicesToken();
+
+            // yield until UnityServicesToken is reset if undefined
+            while (string.IsNullOrEmpty(UnityServicesToken))
+            {
+                await Task.Yield();
+            }
+
             headers.AddAuthorization(UnityServicesToken, ServiceHeaderUtils.k_BearerScheme);
             await Task.CompletedTask;
 #else
@@ -223,34 +289,68 @@ namespace Unity.Cloud.Identity.Editor
 #endif
         }
 
+        void ValidateOrRevokeUnityServicesToken()
+        {
+            // Revoke any existing unity services token if expiry time is less than 30 seconds
+            if (string.IsNullOrEmpty(UnityServicesToken) || m_TokenExpiry == null) return;
+            var dif = (DateTime)m_TokenExpiry - DateTime.UtcNow;
+            // When Unity Hub is running in background, it refreshes the access token 60 seconds before its expiry time
+            // If the unity token is still not refreshed 30 seconds before its expiry time, we assume he Unity Hub is not running
+            // We manually revoke the unity token and activate the Unity Hub to initiate the access token refresh.
+            if ((dif.TotalSeconds > 30)) return;
+            Debug.Log("Revoking Unity services token. Awaiting the Unity Hub to refresh the Unity Editor cloud project settings.");
+            RevokeAndForceRefreshUnityServicesToken();
+        }
+
+        void RevokeAndForceRefreshUnityServicesToken()
+        {
+            UnityServicesToken = string.Empty;
+            m_TokenExpiry = null;
+
+            // Activation of the Hub to initiate the token refresh
+            CloudProjectSettings.RefreshAccessToken(OnRefreshAccessToken);
+            Application.OpenURL($"{k_UnityHubUriScheme}{k_UnityHubLoginDomain}");
+        }
+
+        DateTime? ConvertTimestamp(int timestamp)
+        {
+            if (timestamp == 0) return null;
+            var dateTimeOffset = DateTimeOffset.FromUnixTimeSeconds(timestamp);
+            return dateTimeOffset.UtcDateTime;
+        }
+
         /// <inheritdoc/>
         public async Task<IUserInfo> GetUserInfoAsync()
         {
-            if (m_AuthenticatedUserSession == null)
-            {
-                InitAuthenticatedUserSession();
-            }
+            ThrowIfInvalidAuthenticatedUserSessionState();
             return await m_AuthenticatedUserSession.GetUserInfoAsync();
         }
 
         /// <inheritdoc/>
         public IAsyncEnumerable<IOrganization> ListOrganizationsAsync(Range range, CancellationToken cancellationToken = default)
         {
-            if (m_AuthenticatedUserSession == null)
-            {
-                InitAuthenticatedUserSession();
-            }
+            ThrowIfInvalidAuthenticatedUserSessionState();
             return m_AuthenticatedUserSession.ListOrganizationsAsync(range, cancellationToken);
         }
 
         /// <inheritdoc/>
         public async Task<IOrganization> GetOrganizationAsync(OrganizationId organizationId)
         {
+            ThrowIfInvalidAuthenticatedUserSessionState();
+            return await m_AuthenticatedUserSession.GetOrganizationAsync(organizationId);
+        }
+
+        void ThrowIfInvalidAuthenticatedUserSessionState()
+        {
+            if (!m_AuthenticationState.Equals(AuthenticationState.LoggedIn))
+            {
+                throw new InvalidOperationException("Cannot reach Cloud service, awaiting user Access Token.");
+            }
+
             if (m_AuthenticatedUserSession == null)
             {
                 InitAuthenticatedUserSession();
             }
-            return await m_AuthenticatedUserSession.GetOrganizationAsync(organizationId);
         }
 
         /// <inheritdoc/>